Remove another set of APIs.
