+++ src/main/java/cloud/filibuster/junit/Assertions.java
@@ -1,11 +1,8 @@
 package cloud.filibuster.junit;
 
-import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedAPIException;
 import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedByHTTPServerException;
 import cloud.filibuster.junit.server.core.FilibusterCore;
-
 import static cloud.filibuster.instrumentation.helpers.Property.getServerBackendCanInvokeDirectlyProperty;
-import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedHelper;
 
 /**
  * Assertions provided by Filibuster for writing conditional, fault-based assertions.
@@ -29,22 +26,6 @@ public class Assertions {
             }
         } else {
             throw new FilibusterUnsupportedByHTTPServerException("wasFaultInjectedOnMethodWherePayloadContains only supported with local server.");
-        }
-    }
-
-    /**
-     * Determine if a fault was injected during the current test execution.
-     *
-     * Only works if the target service is instrumented using a server instrumentor and reports its name.
-     *
-     * @param serviceName service name, as reported by a server instrumentor.
-     * @return was fault injected
-     */
-    public static boolean wasFaultInjectedOnService(String serviceName) {
-        if (getServerBackendCanInvokeDirectlyProperty()) {
-            throw new FilibusterUnsupportedAPIException("This API is currently not supported. If applicable, please import the GRPC variant of this method instead.");
-        } else {
-            return wasFaultInjectedHelper("/filibuster/fault-injected/service/" + serviceName);
         }
     }
 }
