+++ src/main/java/cloud/filibuster/junit/assertions/HttpAssertions.java
@@ -1,0 +1,21 @@
+package cloud.filibuster.junit.assertions;
+
+import cloud.filibuster.junit.server.core.FilibusterCore;
+
+import static cloud.filibuster.instrumentation.helpers.Property.getServerBackendCanInvokeDirectlyProperty;
+import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedHelper;
+
+public class HttpAssertions {
+    // TODO
+    public static boolean wasFaultInjectedOnService(String serviceName) {
+        if (getServerBackendCanInvokeDirectlyProperty()) {
+            if (FilibusterCore.hasCurrentInstance()) {
+                return FilibusterCore.getCurrentInstance().wasFaultInjectedOnService(serviceName);
+            } else {
+                return false;
+            }
+        } else {
+            return wasFaultInjectedHelper("/filibuster/fault-injected/service/" + serviceName);
+        }
+    }
+}
