+++ src/test/java/cloud/filibuster/functional/java/basic/WorldTest.java
@@ -1,12 +1,8 @@
-package cloud.filibuster.functional.java.http;
+package cloud.filibuster.functional.java.basic;
 
-import cloud.filibuster.examples.APIServiceGrpc;
-import cloud.filibuster.examples.Hello;
 import cloud.filibuster.instrumentation.helpers.Networking;
-import cloud.filibuster.junit.TestWithFilibuster;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import io.grpc.StatusRuntimeException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 
@@ -23,10 +19,6 @@ import static cloud.filibuster.integration.instrumentation.TestHelper.stopAPISer
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopExternalServerAndWaitUntilUnavailable;
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopHelloServerAndWaitUntilUnavailable;
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopWorldServerAndWaitUntilUnavailable;
-import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class WorldTest {
     public static ManagedChannel apiChannel;
