+++ src/test/java/cloud/filibuster/functional/java/dynamodb/JUnitDynamoDBFilibusterListTablesTest.java
@@ -1,6 +1,5 @@
 package cloud.filibuster.functional.java.dynamodb;
 
-import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedAPIException;
 import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.DynamoDBClientService;
@@ -39,13 +38,9 @@ import java.util.Map;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
-import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class JUnitDynamoDBFilibusterListTablesTest extends JUnitAnnotationBaseTest {
@@ -149,7 +144,6 @@ public class JUnitDynamoDBFilibusterListTablesTest extends JUnitAnnotationBaseTe
                     List<String> methodNames = mapEntry.getValue();
 
                     if (methodNames.stream().anyMatch(method -> wasFaultInjectedOnJavaClassAndMethod(className + "/" + method))) {
-                        assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService(className), "Expected FilibusterUnsupportedAPIException to be thrown");
                         injectedMethodFound = true;
                         break;
                     }
