+++ src/test/java/cloud/filibuster/functional/java/postgresql/JUnitCockroachDBInterceptorTest.java
@@ -1,6 +1,5 @@
 package cloud.filibuster.functional.java.postgresql;
 
-import cloud.filibuster.exceptions.filibuster.FilibusterUnsupportedAPIException;
 import cloud.filibuster.functional.java.JUnitAnnotationBaseTest;
 import cloud.filibuster.instrumentation.libraries.dynamic.proxy.DynamicProxyInterceptor;
 import cloud.filibuster.integration.examples.armeria.grpc.test_services.postgresql.CockroachClientService;
@@ -23,11 +22,9 @@ import java.util.HashSet;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
-import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjectedOnJavaClassAndMethod;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
@@ -65,7 +62,6 @@ public class JUnitCockroachDBInterceptorTest extends JUnitAnnotationBaseTest {
             testExceptionsThrown.add(t.getMessage());
 
             assertTrue(wasFaultInjected(), "An exception was thrown although no fault was injected: " + t);
-            assertThrows(FilibusterUnsupportedAPIException.class, () -> wasFaultInjectedOnService("java.sql.Connection"), "Expected FilibusterUnsupportedAPIException to be thrown: " + t);
             assertTrue(wasFaultInjectedOnJavaClassAndMethod("java.sql.Connection/getSchema") || wasFaultInjectedOnJavaClassAndMethod("java.sql.Driver/connect"), "Fault was not injected on the expected method: " + t);
             assertTrue(t instanceof PSQLException, "Fault was not of the correct type: " + t);
         }
