+++ src/test/java/cloud/filibuster/functional/python/extended/JUnitFilibusterWorldExtendedHttpTest.java
@@ -26,7 +26,7 @@ import java.util.List;
 import java.util.Set;
 
 import static cloud.filibuster.junit.assertions.GenericAssertions.wasFaultInjected;
-import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnService;
+import static cloud.filibuster.junit.assertions.GrpcAssertions.wasFaultInjectedOnService;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -60,7 +60,7 @@ public class JUnitFilibusterWorldExtendedHttpTest extends JUnitBaseTest {
             if (wasFaultInjected()) {
                 testErrorCodesReceived.add(statusCode);
                 assertTrue(validErrorCodes.contains(statusCode));
-                assertTrue(wasFaultInjectedOnService("world"));
+                assertTrue(wasFaultInjectedOnService("World"));
             } else {
                 assertEquals("200", statusCode);
             }
